refactor: move all dashboard settings to settings/dashboard
