+++ src/components/settings/Dashboard/Mobile.vue
@@ -2,11 +2,8 @@
     <v-card flat>
         <v-card-text>
             <v-row>
-                <v-col class="col-12 col-md-6">
-                    <settings-dashboard-sortable viewport-name="desktop" :column="1" />
-                </v-col>
-                <v-col class="col-12 col-md-6">
-                    <settings-dashboard-sortable viewport-name="tablet" :column="2" />
+                <v-col>
+                    <settings-dashboard-sortable viewport-name="mobile" :column="0" />
                 </v-col>
             </v-row>
             <v-row>
@@ -22,16 +19,16 @@
 import Component from 'vue-class-component'
 import { Mixins } from 'vue-property-decorator'
 import DashboardMixin from '@/components/mixins/dashboard'
-import SettingsDashboardSortable from '@/components/inputs/SettingsDashboardSortable.vue'
+import SettingsDashboardSortable from '@/components/settings/Dashboard/Sortable.vue'
+
 @Component({
     components: {
         SettingsDashboardSortable,
     },
 })
-export default class SettingsDashboardTabTablet extends Mixins(DashboardMixin) {
+export default class SettingsDashboardTabMobile extends Mixins(DashboardMixin) {
     resetLayout() {
-        this.$store.dispatch('gui/resetLayout', 'tabletLayout1')
-        this.$store.dispatch('gui/resetLayout', 'tabletLayout2')
+        this.$store.dispatch('gui/resetLayout', 'mobileLayout')
     }
 }
 </script>
