+++ src/components/settings/Dashboard/Tablet.vue
@@ -6,7 +6,7 @@
                     <settings-dashboard-sortable viewport-name="desktop" :column="1" />
                 </v-col>
                 <v-col class="col-12 col-md-6">
-                    <settings-dashboard-sortable viewport-name="desktop" :column="2" />
+                    <settings-dashboard-sortable viewport-name="tablet" :column="2" />
                 </v-col>
             </v-row>
             <v-row>
@@ -21,19 +21,17 @@
 <script lang="ts">
 import Component from 'vue-class-component'
 import { Mixins } from 'vue-property-decorator'
-import draggable from 'vuedraggable'
 import DashboardMixin from '@/components/mixins/dashboard'
-import SettingsDashboardSortable from '@/components/inputs/SettingsDashboardSortable.vue'
-
+import SettingsDashboardSortable from '@/components/settings/Dashboard/Sortable.vue'
 @Component({
     components: {
         SettingsDashboardSortable,
     },
 })
-export default class SettingsDashboardTabDesktop extends Mixins(DashboardMixin) {
+export default class SettingsDashboardTabTablet extends Mixins(DashboardMixin) {
     resetLayout() {
-        this.$store.dispatch('gui/resetLayout', 'desktopLayout1')
-        this.$store.dispatch('gui/resetLayout', 'desktopLayout2')
+        this.$store.dispatch('gui/resetLayout', 'tabletLayout1')
+        this.$store.dispatch('gui/resetLayout', 'tabletLayout2')
     }
 }
 </script>
